+++ src/pages/Main.js
@@ -1,16 +1,13 @@
 import React from 'react';
 import Location from '../components/Location'
 import AddLocation from '../components/AddLocation'
-import {LocationContextProvider} from '../contexts/locationContext'
 
 function Main() {
   return (
-    <LocationContextProvider>
       <>
           <AddLocation />
           <Location />
       </>
-    </LocationContextProvider>
   );
 }
